+++ crates/wdl-ast/src/v1/task/common/container/value.rs
@@ -309,7 +309,6 @@ mod tests {
             .next()
             .expect("the 'hello' task to exist")
             .requirements()
-            .next()
             .expect("the 'requirements' block to exist")
             .items()
             .find_map(|p| p.into_container())